Lox parser for chapter 6 of the book implemented with test coverage. (#2)
**Summary**
I'm only through chapter 6 of the book Crafting Interpreters. This is a Rust based parser for expressions in the Lox language, which is the only component of the language we have so far.

There is decent unit test coverage for turning the tokens into an AST. There is only one integration test for now as I think about the best way of reading in source files and making sure they equal the expected tokens and/or AST.

There are many small changes across the codebase because I added cargo clippy as static analysis.

**Changes**
- Use purpose built location struct which has the filename in addition to the line number.
- Use a general lox error struct which covers errors from all "phases". Lexing, parsing, and executing for the interpreter.
- Use C#-ish syntax to build the lexed tokens to AST parser.

**Issues**
There is diff noise because I changed a few file names and got better at extracting from the "one giant file" that was the scanner/lexer.