Fix: Lexer/scanner bug when scanning identifiers. (#4)
When scanning an identifier there is a bug where the iterator "eats" the following character, does not include it in the identifier string (which is correct), and discards it (not correct). This causes problems anytime people don't put a space after their identifier or keyword. This is very noticeable when doing function declarations like
```
fun my_function_name() {}
```
This results in the following tokens
1. Keyword - `fun` (this is the function token)
2. Identifier - `my_function_name`
3. Right parenthesis - `)`
4. Left brace - `{`
5. Right brace - `}`

The bug is the missing the left parenthesis token between 2 and 3.

While fixing the bug I introduced a new way of getting the token's lexeme that let me simplify the code for scanning other types of tokens.