Feat: Add statements and global variables (#6)
- Wire up statements. 
- All expression or print statements must end with a semicolon `;`

**Variables**
- Variables get saved and their values can be printed!
- Support for re-assigning to variables.
- Implement variable assignment.


**Environment for variables**
The Lox variable environment is passed as a separate parameter to make its mutable state more clear. Instead of the book implementation I use a vector of scopes behind the scenes instead of recursive scopes. This works much better with Rust and avoids all of the lifetime and ownership issues of nested mutable dictionaries.

There are tests for the environment API. These should pass even if the internals change.

Co-authored-by: Joshua Rakita <joraki@microsoft.com>